requirements.txt
Pillow>=8.0.1
PyQt5>=5.15.1
ffmpeg-python>=0.2.0
matplotlib>=3.3.3
tqdm>=4.51.0
